requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.4.2
seaborn>=0.11.1
tqdm>=4.61.2
PyYAML>=5.4.1
nltk>=3.6.2
rouge-score>=0.0.4
scikit-learn>=0.24.2
transformers>=4.9.2